% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTissueMask.R
\name{calcTissueMaskParallel}
\alias{calcTissueMaskParallel}
\title{Calculates the tissue mask of a given image}
\usage{
calcTissueMaskParallel(
  image,
  tissue_index = NULL,
  size_selection = 10,
  BPPARAM = BPPARAM
)
}
\arguments{
\item{image}{A multiplexed image or list of multiplexed images, can be a CytoImageList}

\item{tissue_index}{a numerical index that corresponds to the tissue markers in the image object.}

\item{BPPARAM}{The number of parallel processing cores to be used}
}
\value{
Tissue masks of given image/s (area of the image/s containing tissue)
}
\description{
Calculates the tissue mask of a given image
}
\examples{

image.mask <- calcTissueMaskParallel(image = image.list, tissue_index = c(2,3,4), cores = 40)

}
