% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCells.R
\name{createCells}
\alias{createCells}
\title{Perform simple segmentation of multiplexed cellular images}
\usage{
createCells(mask, image)
}
\arguments{
\item{mask}{A mask or list of masks, can be a CytoImageList corresponding to the given image/s}

\item{image}{An image or list of images, can be a CytoImageList}
}
\value{
A matrix with individual cells as rows and features: mean marker intensities, x/y coordinates, area and image number as columns
}
\description{
Perform simple segmentation of multiplexed cellular images
}
\examples{

cell.matrix <- createCells(mask.list, image.list)

}
