% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cells.R
\name{normalizeCells}
\alias{normalizeCells}
\title{Normalizes and transforms cell data in preparation for clustering (accepts dataframes, singleCellExperiments and SpatialExperiments)}
\usage{
normalizeCells(
  cells,
  markers = NULL,
  assayIn = NULL,
  assayOut = "norm",
  imageID = "imageID",
  transformation = NULL,
  method = NULL,
  cores = 1
)
}
\arguments{
\item{cells}{A Dataframe of singlecellexperement or spatialexperiment containing cells and features to be normalized/transformed}

\item{markers}{A list containing the names of cell markers which will be normalized/transformed}

\item{assayIn}{If input is a SCE or SE with multiple assays, specify the assay to be normalized/transformed}

\item{assayOut}{If input is a SCE or SE, the new of the normalized data.}

\item{imageID}{If input is a SCE or SE, this is the name of the image ID variable in order to stratify cells correctly}

\item{transformation}{The transformation/s to be performed, default is NULL, accepted values: 'asinh', 'sqrt', 'log'}

\item{method}{The normalization method/s to be performed, default is NULL, accepted values: 'meandiv', '99perc', '1stPC'}
}
\value{
returns a dataframe with individual cells as rows and features as columns
}
\description{
Normalizes and transforms cell data in preparation for clustering (accepts dataframes, singleCellExperiments and SpatialExperiments)
}
\examples{

cells.normalized <- normalizeCells(cells = cells.sce, markers = c('SMA', 'CD44', 'CD45', 'cyt-19') isSCE = TRUE, assayName = 'cells', imageNb = 'ImageID', transformation = 'asinh', method = '99perc')

}
