% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSeg.R
\name{simpleSeg}
\alias{simpleSeg}
\title{Perform simple segmentation of multiplexed cellular images}
\usage{
simpleSeg(
  image,
  nucleus = 1,
  cellBody = "dilate",
  sizeSelection = 10,
  smooth = 1,
  transform = c("norm99, maxThresh, asin"),
  tolerance = 0.01,
  ext = 1,
  discSize = 3,
  cores = 5
)
}
\arguments{
\item{image}{An image or list of images or cytoimagelist to be read into the function.}

\item{nucleus}{the channel number of the nuclei marker}

\item{cellBody}{method of cytoplasm identification. Can be "dilate", "diskModel" or "markerModel"}

\item{sizeSelection}{minimum pixels for an object to be recognised as signal and not noise}

\item{smooth}{the amount of smoothing to be applied to the nuclei marker channle}

\item{transform}{a list containing desired normalization/transformation methods to be performed prior to cytoplasm identification, accepted values are 'mixMax' and/or 'asin'}

\item{tolerance}{The minimum height of the object in the units of image intensity between its highest point (seed) and the point where it contacts another object (checked for every contact pixel). If the height is smaller than the tolerance, the object will be combined with one of its neighbors, which is the highest. Tolerance should be chosen according to the range of x. Default value is 1, which is a reasonable value if x comes from distmap.}

\item{ext}{Radius of the neighborhood in pixels for the detection of neighboring objects. Higher value smoothes out small objects.}

\item{discSize}{size of dilation around nuclei to create cell disk #dilation size}

\item{cores}{= 50 number or cores for paralell processing}

\item{BPPARAM}{A BiocParallelParam object.}

\item{norm99perfrom}{99th percentile transformation}

\item{maxThresh}{scale intensities between 0 and 1}

\item{nucNormalize}{a list containing desired normalization/transformation methods to be performed prior to nucleus identification, accepted values are 'max Thresh' and 'norm99perform'}

\item{minMax}{scale image channel intensities between 0 and 1}

\item{asin}{perform asinh normalization on image channels}

\item{cyt_index}{index of the cytoplasm marker channel. Use if cellBody = markerModel}
}
\value{
A list of image masks
}
\description{
Perform simple segmentation of multiplexed cellular images
}
\examples{

1+1

}
