% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSeg.R
\name{simpleSeg}
\alias{simpleSeg}
\title{Perform simple segmentation}
\usage{
simpleSeg(image)
}
\arguments{
\item{image}{An image}

\item{BPPARAM}{A BiocParallelParam object.}
}
\value{
A list of image masks
}
\description{
Perform simple segmentation
}
\examples{

1+1

}
